backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "backend "
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.9",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.3.1",
    "django-cors-headers>=4.3.1",
    "psycopg2-binary>=2.9.11",
    "dj-database-url>=2.2.0",
    "gunicorn>=22.0.0",
]
